Remove `backend_or_name` in favor of `platforms`.
Throughout the lowering code we pass `backend_or_name`, which
can be a XlaBackend object, a platform name, or None. At the
same time we pass `platforms` which can be a sequence of
platform names.

There is already code to create an `XlaBackend` from `platforms` (only for the case when `platforms` is a singleton, using `xla_bridge.get_backend`).

PiperOrigin-RevId: 706127161